adding works

+++ frontend/src/app/components/categories/edit-category/edit-category.component.ts
@@ -4,7 +4,6 @@ import {CategoriesService} from "../../../services/categories.service";
 import {ActivatedRoute, Router} from "@angular/router";
 import {Category} from "../../../model/category";
 import {HttpResponse} from "@angular/common/http";
-import {Product} from "../../../model/Product";
 import {Observable} from "rxjs";
 
 @Component({
@@ -26,6 +25,25 @@ export class EditCategoryComponent implements OnInit {
     newFieldsTypes: this.fb.array([]),
     oldFieldsIds: this.fb.array([])
   })
+
+  ngOnInit(): void {
+    this.getCategories();
+    this.route.paramMap.subscribe((params) =>
+    {this.uuid = params.get('uuid')!.toString()});
+    this.categoriesService.getCategory(this.uuid).subscribe(
+      (data: HttpResponse<Category>) => {
+        this.category = data.body!;
+        for(let i = 0;i < this.category.fields.length; i++){
+          this.getOldFieldsNames.push(this.fb.control(this.category.fields[i].name));
+          this.getOldFieldsTypes.push(this.fb.control(this.category.fields[i].type));
+          this.getOldFieldsIds.push(this.fb.control(this.category.fields[i].fieldID));
+        }
+        this.editCategoryForm.patchValue({
+          name: this.category.name,
+          parent: this.category.parentCategory?.categoryID
+        })
+      });
+  }
 
   getOldFieldName(i : number) {
     return this.getOldFieldsNames.at(i);
@@ -58,7 +76,7 @@ export class EditCategoryComponent implements OnInit {
     this.getNewFieldsNames.push(new FormControl('', [Validators.required]));
   }
   addFieldsTypes() {
-    this.getNewFieldsTypes.push(new FormControl('', [Validators.required]));
+    this.getNewFieldsTypes.push(new FormControl('Int', [Validators.required]));
   }
   private removeOldFieldsNames(i : number) {
     this.getOldFieldsNames.removeAt(i);
@@ -93,60 +111,45 @@ export class EditCategoryComponent implements OnInit {
 
   ) {}
 
-  ngOnInit(): void {
-    this.getCategories();
-    this.route.paramMap.subscribe((params) =>
-    {this.uuid = params.get('uuid')!.toString()});
-    this.categoriesService.getCategory(this.uuid).subscribe(
-      (data: HttpResponse<Category>) => {
-      this.category = data.body!;
-      for(let i = 0;i < this.category.fields.length; i++){
-        this.getOldFieldsNames.push(this.fb.control(this.category.fields[i].name));
-        this.getOldFieldsTypes.push(this.fb.control(this.category.fields[i].type));
-        this.getOldFieldsIds.push(this.fb.control(this.category.fields[i].fieldID));
-      }
-      this.editCategoryForm.patchValue({
-        name: this.category.name,
-        parent: this.category.parentCategory?.categoryID
-      })
-    });
-  }
+
   removeNewFields(i : number){
     this.removeNewFieldsNames(i);
     this.removeNewFieldsTypes(i);
   }
   removeOldFields(i : number){
-    this.categoriesService.removeField(this.getOldFieldsIds.getRawValue()[i])
-    .subscribe((result) => {
-      if (result.status === 200) {}})
-    this.removeOldFieldsNames(i);
-    this.removeOldFieldsTypes(i);
-    this.getOldFieldsIds.removeAt(i);
-
+    if(confirm("Do you really want to delete this field? This action can't be undone")){
+      this.categoriesService.removeField(this.getOldFieldsIds.getRawValue()[i])
+        .subscribe((result) => {
+          if (result.status === 200) {
+            this.removeOldFieldsNames(i);
+            this.removeOldFieldsTypes(i);
+            this.getOldFieldsIds.removeAt(i);
+          }})
+    }
   }
   editCategory() {
     if (this.editCategoryForm.valid) {
 
       for(let i = 0;i < this.getOldFieldsTypes.getRawValue().length; i++){
-          const FieldDTO: object = {
-            "name": this.getOldFieldsNames.getRawValue()[i],
-            "type": this.getOldFieldsTypes.getRawValue()[i]
-          };
-          console.log(this.getOldFieldsIds.getRawValue()[i]);
-          this.categoriesService.updateField(this.getOldFieldsIds.getRawValue()[i], FieldDTO)
-            .subscribe((result) => {
-                if (result.status === 200) {}})}
-      /*
+        const FieldDTO: object = {
+          "name": this.getOldFieldsNames.getRawValue()[i],
+          "type": this.getOldFieldsTypes.getRawValue()[i]
+        };
+        console.log(this.getOldFieldsIds.getRawValue()[i]);
+        this.categoriesService.updateField(this.getOldFieldsIds.getRawValue()[i], FieldDTO)
+          .subscribe((result) => {
+            if (result.status === 200) {}})}
+
       for(let i = 0;i < this.getNewFieldsTypes.getRawValue().length; i++){
         const FieldDTO: object = {
-          "name": this.getNewFieldsNames.getRawValue()[0],
-          "type": this.getNewFieldsTypes.getRawValue()[0]
+          "name": this.getNewFieldsNames.getRawValue()[i],
+          "type": this.getNewFieldsTypes.getRawValue()[i]
         };
         this.categoriesService.addField(this.uuid, FieldDTO)
           .subscribe((result) => {if (result.status === 200) {
-            }})
+          }})
       }
-      */
+
       const CategoryDTO: object = {
         "name": this.editCategoryForm.getRawValue().name,
         "parentCategoryID": this.editCategoryForm.getRawValue().parent
